created the ScanPath class from BeamCenter and redistributed functionality from HeatSource
